[DREALODEMA-128] refactoring compute data from technical sheet page to children components
